clean up

+++ server/controller/house/house.js
@@ -1,4 +1,5 @@
 const createHouseService = require("../../services/house/house");
+const House = require("../../models/houses/house");;
 
 const CreateHouse = async (req, res) => {
   try {
@@ -68,6 +69,11 @@ const GetHouseById = async (req, res) => {
 const DeleteHouse = async (req, res) => {
   try {
     const houseId = req.params.houseId;
+    const userId = req.body.userId;
+    const isAuthorized = await CheckHouseOwnership(houseId, userId);
+    if (!isAuthorized) {
+      return res.status(403).json({ error: "You are not authorized to delete this house" });
+    }
     await createHouseService.DeleteHouse(houseId);
     return res.status(200).json({ message: "House deleted successfully" });
   } catch (error) {
@@ -80,6 +86,13 @@ const EditHouse = async (req, res) => {
   try {
     const houseId = req.params.houseId;
     const updatedData = req.body;
+    const userId = req.body.userId;
+    console.log('userId', userId)
+    const isAuthorized = await CheckHouseOwnership(houseId, userId);
+    console.log("a", isAuthorized)
+    if (!isAuthorized) {
+      return res.status(403).json({ error: "You are not authorized to delete this house" });
+    }
     const updatedHouse = await createHouseService.EditHouse(
       houseId,
       updatedData
@@ -88,6 +101,19 @@ const EditHouse = async (req, res) => {
   } catch (error) {
     console.log(error);
     return res.status(500).json({ error: "Internal Server Error" });
+  }
+};
+
+const CheckHouseOwnership = async (houseId, userId) => {
+  try {
+    const house = await House.findOne({ where: { id: houseId } });
+    console.log("house:", house); // Debugging
+    if (house && house.userId === userId) {
+      return true;
+    }
+    return false;
+  } catch (error) {
+    throw error;
   }
 };
